Add typing stubs
This is copy/pasted from the stubs of the `py.iniconfig` module in the
`py` library, however I removed the `_` prefix from `SectionWrapper`
type since it's good to be able to refer to it. I marked its `__init__`
as private (there is unfortunately no programmatic way to mark this).

Tested working against pytest's code base.

Co-authored-by: Daniel Hahler <github@thequod.de>